requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
deep-translator==1.11.4
Django==5.1.1
django-autotranslate==1.3.0
django-cloudinary-storage==0.3.0
django-cockroachdb==5.1
django-deep-translator==1.5.1
futures==3.0.5
goslate==1.5.4
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
polib==1.2.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
whitenoise==6.7.0
